test: menyesuaikan unittest dengan perubahan.
