--- tests/SastrawiTest.pyproj
@@ -1,73 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<Project ToolsVersion="4.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003" DefaultTargets="Build">
-  <PropertyGroup>
-    <Configuration Condition=" '$(Configuration)' == '' ">Debug</Configuration>
-    <SchemaVersion>2.0</SchemaVersion>
-    <ProjectGuid>{69199be5-44c5-45c3-8b82-62f14da2b9f1}</ProjectGuid>
-    <ProjectHome />
-    <StartupFile />
-    <SearchPath>..\src\</SearchPath>
-    <WorkingDirectory>.</WorkingDirectory>
-    <OutputPath>.</OutputPath>
-    <ProjectTypeGuids>{888888a0-9f3d-457c-b088-3a5042f75d52}</ProjectTypeGuids>
-    <LaunchProvider>Standard Python launcher</LaunchProvider>
-    <InterpreterId>{9a7a9026-48c1-4688-9d5d-e5699d47d074}</InterpreterId>
-    <InterpreterVersion>2.7</InterpreterVersion>
-    <Name>SastrawiTest</Name>
-  </PropertyGroup>
-  <PropertyGroup Condition="'$(Configuration)' == 'Debug'" />
-  <PropertyGroup Condition="'$(Configuration)' == 'Release'" />
-  <PropertyGroup>
-    <VisualStudioVersion Condition=" '$(VisualStudioVersion)' == '' ">10.0</VisualStudioVersion>
-    <PtvsTargetsFile>$(MSBuildExtensionsPath32)\Microsoft\VisualStudio\v$(VisualStudioVersion)\Python Tools\Microsoft.PythonTools.targets</PtvsTargetsFile>
-  </PropertyGroup>
-  <ItemGroup>
-    <InterpreterReference Include="{9a7a9026-48c1-4688-9d5d-e5699d47d074}\2.7" />
-  </ItemGroup>
-  <ItemGroup>
-    <ProjectReference Include="..\src\Sastrawi\Sastrawi.pyproj">
-      <Name>Sastrawi</Name>
-      <Project>{9a2db5b1-a5b0-4b52-aa46-d2ce718cc5dd}</Project>
-      <Private>True</Private>
-    </ProjectReference>
-  </ItemGroup>
-  <ItemGroup>
-    <Folder Include="FunctionalTests\" />
-    <Folder Include="FunctionalTests\Stemmer\" />
-    <Folder Include="IntegrationTests\" />
-    <Folder Include="IntegrationTests\Stemmer\" />
-    <Folder Include="UnitTests\" />
-    <Folder Include="UnitTests\Dictionary\" />
-    <Folder Include="UnitTests\Morphology\" />
-    <Folder Include="UnitTests\Morphology\Disambiguator\" />
-    <Folder Include="UnitTests\StopWordRemover\" />
-    <Folder Include="UnitTests\Stemmer\" />
-  </ItemGroup>
-  <ItemGroup>
-    <Compile Include="FunctionalTests\Stemmer\stemmer_test.py" />
-    <Compile Include="FunctionalTests\Stemmer\__init__.py">
-      <SubType>Code</SubType>
-    </Compile>
-    <Compile Include="FunctionalTests\__init__.py">
-      <SubType>Code</SubType>
-    </Compile>
-    <Compile Include="IntegrationTests\Stemmer\stemmer_test.py" />
-    <Compile Include="IntegrationTests\Stemmer\__init__.py" />
-    <Compile Include="IntegrationTests\__init__.py" />
-    <Compile Include="UnitTests\Dictionary\array_dictionary_test.py" />
-    <Compile Include="UnitTests\Dictionary\__init__.py" />
-    <Compile Include="UnitTests\Morphology\Disambiguator\disambiguator_prefix_rule1_test.py" />
-    <Compile Include="UnitTests\Morphology\Disambiguator\__init__.py" />
-    <Compile Include="UnitTests\Morphology\invalid_affix_pair_specification_test.py" />
-    <Compile Include="UnitTests\Morphology\__init__.py" />
-    <Compile Include="UnitTests\Stemmer\stemmer_factory_test.py" />
-    <Compile Include="UnitTests\Stemmer\stemmer_test.py" />
-    <Compile Include="UnitTests\Stemmer\__init__.py" />
-    <Compile Include="UnitTests\StopWordRemover\stop_word_remover_factory_test.py" />
-    <Compile Include="UnitTests\StopWordRemover\stop_word_remover_test.py" />
-    <Compile Include="UnitTests\StopWordRemover\__init__.py" />
-    <Compile Include="UnitTests\__init__.py" />
-  </ItemGroup>
-  <Import Project="$(PtvsTargetsFile)" Condition="Exists($(PtvsTargetsFile))" />
-  <Import Project="$(MSBuildToolsPath)\Microsoft.Common.targets" Condition="!Exists($(PtvsTargetsFile))" />
-</Project>
